Rename 'garbage_collection' to 'garbage_collector'
