+++ src/include/garbage_collector.h
